adapt polling
